rename/remove: mockdata.json을 control 폴더로 옮김, tabForList 폴더 이름을 tabForDispatchLis로 변경
